third_party/WebKit/LayoutTests/media/track/track-cues-pause-on-exit.html

<!DOCTYPE html>
<title>Tests that the video is paused after cues that have pause-on-exit flag are processed.</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../media-file.js"></script>
<video controls>
    <track src="captions-webvtt/simple-captions.vtt" default>
</video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    var track = document.querySelector("track");
    track.onload = t.step_func(function() {
        assert_equals(track.track.cues.length, 4);
        for (var i = 0; i < track.track.cues.length; ++i) {
            var cue = track.track.cues[i];
            if (i % 2 == 0) {
                cue.pauseOnExit = true;
                cue.onexit = t.step_func(function(event) {
                    assert_true(video.paused);

                    video.play();

                    if (event.target.id == 2)
                        t.done();
                });
            }
        }
        video.src = findMediaFile("video", "../content/test");
        video.currentTime = 4.00;
        video.play();
        assert_false(video.paused);
    });
});
</script>
